pyproject.toml
[tool.poetry]
name = "vacansee-bot"
version = "0.1.0"
description = ""
authors = ["Baclrary <ihorprotsak@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
httpx = "^0.26.0"
python-telegram-bot = "^20.8"
asyncpg = "^0.29.0"
sqlalchemy = "^2.0.27"
alembic = "^1.13.1"
python-dotenv = "^1.0.1"
aiofiles = "^23.2.1"
beautifulsoup4 = "^4.12.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
